requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==4.2.2
annotated-types==0.7.0
anyio==4.4.0
arrow==1.3.0
async-timeout==4.0.3
attrs==24.2.0
audio-recorder-streamlit==0.0.10
binaryornot==0.4.4
blinker==1.9.0
cachetools==5.5.0
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cookiecutter==2.6.0
cycler==0.12.1
distlib==0.3.4
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
entrypoints==0.4
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-cli==0.0.4
ffmpy==0.4.0
filelock==3.4.2
flet==0.22.0
flet-core==0.22.0
flet-runtime==0.22.0
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.10
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.2
googleapis-common-protos==1.65.0
gradio==4.44.1
gradio_client==1.3.0
grpcio==1.66.2
grpcio-status==1.66.2
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.25.1
idna==3.7
importlib_resources==6.4.5
Jinja2==3.1.4
jiter==0.6.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
meteostat==1.6.8
multidict==6.1.0
narwhals==1.25.2
numpy==2.1.1
oauthlib==3.2.2
openai==0.28.0
orjson==3.10.7
packaging==23.2
pandas==2.2.3
pillow==10.4.0
platformdirs==2.4.1
propcache==0.2.0
proto-plus==1.24.0
protobuf==5.28.2
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pyjokes==0.6.0
pyparsing==3.1.4
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.1
qrcode==7.4.2
referencing==0.36.2
repath==0.9.0
requests==2.32.3
rich==13.7.1
rpds-py==0.22.3
rsa==4.9
ruff==0.6.9
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.37.2
streamlit==1.42.0
tenacity==9.0.0
text-unidecode==1.3
toml==0.10.2
tomlkit==0.12.0
toolz==1.0.0
tornado==6.4.2
tqdm==4.66.5
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.1
virtualenv==20.13.0
watchdog==4.0.1
watchfiles==0.22.0
websockets==12.0
yarl==1.15.0
